Island: Move IslandModeService out of unnecessary subpackage
