+++ monkey/monkey_island/cc/services/__init__.py
@@ -2,4 +2,4 @@ from .authentication.authentication_service import AuthenticationService
 from .authentication.json_file_user_datastore import JsonFileUserDatastore
 
 from .aws import AWSService
-from .mode.island_mode_service import IslandModeService
+from .island_mode_service import IslandModeService
