+++ monkey/monkey_island/cc/services/island_mode_service.py
